+++ src/extension.ts
@@ -1,11 +1,11 @@
 import * as vscode from 'vscode';
-import { TitleParser } from './titleParser';
+import { DocParser } from './docParser';
 import { DocFxItem } from './docfx-item';
 import { docFxItemsListFactory } from './docfx-items-list';
-import { SortingType } from './docfx-items-sorting.enum';
+import { SortingTypeOptions } from './configurations';
 
 export function activate(context: vscode.ExtensionContext) {
-	let disposable = vscode.commands.registerCommand('doxfx-toc-generator.generateYamlToc', async (folder: vscode.Uri) => {
+	let disposable = vscode.commands.registerCommand('docfx-toc-generator.generateYamlToc', async (folder: vscode.Uri) => {
 
 		if (folder instanceof vscode.Uri === false) {
 			vscode.window.showErrorMessage("Invalid command execution context");
@@ -17,11 +17,11 @@ export function activate(context: vscode.ExtensionContext) {
 			vscode.window.showWarningMessage(`No files found under ${folder.fsPath}`);
 		}
 
-		const docFxItems = docFxItemsListFactory(SortingType.fileName, true);
+		const docFxItems = docFxItemsListFactory(SortingTypeOptions.fileName);
 
 		for (const file of files) {
 			const doc = await vscode.workspace.openTextDocument(file);
-			const titleParser = new TitleParser(doc);
+			const titleParser = new DocParser(doc);
 			let title = titleParser.extractTitle();
 			if (!title) {
 				vscode.window.showWarningMessage(`No title found in ${file.fsPath}`);