+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Change Log
 
-All notable changes to the "doxfx-toc-generator" extension will be documented in this file.
+All notable changes to the "docfx-toc-generator" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [0.0.2] - - 2021-07-21
+## [0.0.3] - 2021-08-15
+
+- Added support to configure sorting ascending or descending.
+
+## [0.0.2] - 2021-07-21
 
 - Refactoring to support different sorting strategies.